# backend/openapi.yaml
openapi: 3.0.3
info:
  title: VoterField API
  description: Backend API for the VoterField Canvassing Platform.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production API

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRole:
      type: string
      enum: [admin, canvasser]

    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [active, suspended, pending_delete] }
        plan_id: { type: string }
        limits: { type: object }
        last_activity_at: { type: string, format: date-time }
      required: [id, name, status]

    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
      required: [id, orgId, name, email, role]

    Voter:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        externalId: { type: string }
        firstName: { type: string }
        middleName: { type: string }
        lastName: { type: string }
        suffix: { type: string }
        age: { type: integer }
        gender: { type: string }
        race: { type: string }
        party: { type: string }
        phone: { type: string }
        address: { type: string }
        unit: { type: string }
        city: { type: string }
        state: { type: string }
        zip: { type: string }
        geom:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        lastInteractionStatus: { type: string }
        lastInteractionTime: { type: string, format: date-time }
      required: [id, orgId, externalId, firstName, lastName, address, city, zip]

    WalkList:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        name: { type: string }
        voterIds:
          type: array
          items: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        createdByUserId: { type: string, format: uuid }
      required: [id, orgId, name, voterIds, createdAt, createdByUserId]

    Assignment:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        listId: { type: string, format: uuid }
        canvasserId: { type: string, format: uuid }
        status: { type: string, enum: [assigned, in_progress, completed] }
        createdAt: { type: string, format: date-time }
      required: [id, orgId, listId, canvasserId, status, createdAt]

    Interaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        voter_id: { type: string, format: uuid }
        assignment_id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time }
        channel: { type: string }
        result_code: { type: string }
        notes: { type: string }
        client_interaction_uuid: { type: string }
        survey_responses: { type: object }
      required: [id, org_id, user_id, voter_id, occurred_at, channel, result_code, client_interaction_uuid]

    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        type: { type: string }
        status: { type: string, enum: [pending, processing, completed, failed] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        result: { type: object }
        error: { type: string }
        metadata: { type: object }

    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]

    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
        org: { $ref: '#/components/schemas/Organization' }

security:
  - BearerAuth: []

paths:
  /auth/login:
    post:
      summary: Authenticate user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /auth/switch-role:
    post:
      summary: Switch current role (dev utility)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/UserRole' }

  /me:
    get:
      summary: Get current user context
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /org:
    get:
      summary: Get current organization context
      responses:
        '200':
          description: Current organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'

  /voters:
    get:
      summary: Get voters (paginated)
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Voter list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Voter' }
    post:
      summary: Create voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Voter'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voter'

  /voters/{id}:
    get:
      summary: Get voter by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Voter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voter'
    patch:
      summary: Update voter
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated

  /lists:
    get:
      summary: Get walk lists
      responses:
        '200':
          description: Walk lists
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WalkList' }
    post:
      summary: Create walk list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                voter_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkList'

  /assignments:
    get:
      summary: Get assignments (scope=me|org)
      parameters:
        - in: query
          name: scope
          schema: { type: string }
      responses:
        '200':
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assignment' }
    post:
      summary: Assign list to canvasser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                list_id: { type: string, format: uuid }
                canvasser_id: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'

  /interactions:
    get:
      summary: Get interactions
      responses:
        '200':
          description: Interactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Interaction' }
    post:
      summary: Log interaction (idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Interaction'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interaction'

  /interactions/bulk:
    post:
      summary: Bulk log interactions (idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Interaction'
      responses:
        '200':
          description: Bulk result

  /users:
    get:
      summary: Get users
      parameters:
        - in: query
          name: role
          schema: { $ref: '#/components/schemas/UserRole' }
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }

  /users/invite:
    post:
      summary: Invite new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                role: { $ref: '#/components/schemas/UserRole' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /jobs/import-voters:
    post:
      summary: Start async voter import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/Voter' }
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

  /imports/voters:
    post:
      summary: Upload CSV and start import
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

  /jobs/{id}:
    get:
      summary: Get job status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

  /metrics/field/summary:
    get:
      summary: Field metrics summary
      responses:
        '200':
          description: Summary stats

  /internal/organizations:
    get:
      summary: List all organizations (Staff Only)
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'

  /internal/organizations/{id}/health:
    get:
      summary: Organization health (Staff Only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Org health summary